Distance between two points

+++ conditionals_loop_03/Medium_Ques/Ques_05.java
@@ -11,7 +11,7 @@ public class Ques_05 {
         System.out.println("Enter the x and y coordinates of both points (pair (x,y) of each point)");
         double x1 = sc.nextDouble(), y1 = sc.nextDouble(), x2 = sc.nextDouble(), y2 = sc.nextDouble();
 
-        System.out.println("The distance between two points is " + ((double)(Math.sqrt(Math.pow((y2-y1), 2) + Math.pow((x2-x1), 2)))));
+        System.out.println("The distance between two points is " + ((Math.sqrt(Math.pow((y2-y1), 2) + Math.pow((x2-x1), 2)))));
 
         sc.close();
     }
